Extract strata-graph crate from engine (Phase 2 of #1406)
Move the graph module (12,851 lines) to a new `strata-graph` crate:
- All graph/*.rs files → crates/graph/src/
- branch_dag.rs → crates/graph/src/branch_dag.rs
- branch_status_cache.rs → crates/graph/src/branch_status_cache.rs
- Branch DAG types (DagBranchStatus, etc.) moved to strata-core

GraphSubsystem implements the Subsystem trait for init_system_branch
and load_status_cache_readonly. Database::open() no longer initializes
the branch DAG — the executor handles it via init_system_branch().

Subsystem trait updated: recover() now takes &Arc<Database> instead
of &Database to support subsystems that need Arc (GraphStore, etc.).

Engine crate shrinks from ~60k to ~47k lines.
