+++ crates/engine/src/primitives/branch/index.rs
@@ -282,7 +282,9 @@ impl BranchIndex {
                 .filter_map(|(k, _)| {
                     let key_str = String::from_utf8(k.user_key.to_vec()).ok()?;
                     // Filter out index keys (legacy) and system branches
-                    if key_str.contains("__idx_") || crate::branch_dag::is_system_branch(&key_str) {
+                    if key_str.contains("__idx_")
+                        || strata_core::branch_dag::is_system_branch(&key_str)
+                    {
                         None
                     } else {
                         Some(key_str)
