+++ crates/executor/src/bridge.rs
@@ -14,10 +14,11 @@ use strata_core::limits::Limits;
 use strata_core::primitives::json::{JsonPath, JsonValue};
 use strata_core::{StrataError, StrataResult, Value};
 use strata_engine::{
-    BranchIndex as PrimitiveBranchIndex, Database, EventLog as PrimitiveEventLog, GraphStore,
+    BranchIndex as PrimitiveBranchIndex, Database, EventLog as PrimitiveEventLog,
     JsonStore as PrimitiveJsonStore, KVStore as PrimitiveKVStore,
     SpaceIndex as PrimitiveSpaceIndex, VectorStore as PrimitiveVectorStore,
 };
+use strata_graph::GraphStore;
 
 use crate::types::BranchId;
