+++ crates/graph/src/integrity.rs
@@ -77,9 +77,9 @@ impl GraphStore {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::database::Database;
-    use crate::graph::types::*;
+    use crate::types::*;
     use std::sync::Arc;
+    use strata_engine::Database;
 
     fn setup() -> (Arc<Database>, GraphStore) {
         let db = Database::cache().unwrap();
