+++ crates/graph/src/ontology.rs
@@ -812,8 +812,8 @@ fn json_type_name(v: &serde_json::Value) -> &'static str {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::database::Database;
     use std::sync::Arc;
+    use strata_engine::Database;
 
     fn setup() -> (Arc<Database>, GraphStore) {
         let db = Database::cache().unwrap();
