+++ crates/graph/src/packed.rs
