+++ crates/graph/src/traversal.rs
@@ -284,8 +284,8 @@ impl GraphStore {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::database::Database;
     use std::sync::Arc;
+    use strata_engine::Database;
 
     fn setup() -> (Arc<Database>, GraphStore) {
         let db = Database::cache().unwrap();
