+++ tests/executor/common.rs
@@ -7,6 +7,7 @@ use strata_executor::{Executor, Output, Session, Strata};
 /// Create an executor with an in-memory database
 pub fn create_executor() -> Executor {
     let db = Database::cache().unwrap();
+    strata_graph::branch_dag::init_system_branch(&db);
     Executor::new(db)
 }
 
@@ -18,12 +19,15 @@ pub fn create_strata() -> Strata {
 /// Create a Session with an in-memory database
 pub fn create_session() -> Session {
     let db = Database::cache().unwrap();
+    strata_graph::branch_dag::init_system_branch(&db);
     Session::new(db)
 }
 
 /// Create a database for shared use
 pub fn create_db() -> Arc<Database> {
-    Database::cache().unwrap()
+    let db = Database::cache().unwrap();
+    strata_graph::branch_dag::init_system_branch(&db);
+    db
 }
 
 /// Helper to create an event payload (must be an Object)
